requirements.txt
Django==1.6.2
South==0.8.4
mysql-python
django-tastypie==0.11.0
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.6.1
wsgiref==0.1.2
